# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OcholaSupernet API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: ispId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerBody"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        ispId:
          type: integer
        name:
          type: string
        username:
          type: string
        phone:
          type: string
        email:
          type: string
        planId:
          type: integer
        planName:
          type: string
        status:
          type: string
          enum: [active, expired, suspended, pending]
        balance:
          type: number
        ipAddress:
          type: string
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - ispId
        - name
        - username
        - status
        - createdAt
    CreateCustomerBody:
      type: object
      properties:
        ispId:
          type: integer
        name:
          type: string
        username:
          type: string
        phone:
          type: string
        email:
          type: string
        planId:
          type: integer
        password:
          type: string
      required:
        - ispId
        - name
        - username
    UpdateCustomerBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        planId:
          type: integer
        status:
          type: string
        balance:
          type: number
...