# lib/api-zod/src/generated/types/createCustomerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OcholaSupernet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCustomerBody {
  ispId: number;
  name: string;
  username: string;
  phone?: string;
  email?: string;
  planId?: number;
  password?: string;
}