# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OcholaSupernet API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerStatus } from "./customerStatus";

export interface Customer {
  id: number;
  ispId: number;
  name: string;
  username: string;
  phone?: string;
  email?: string;
  planId?: number;
  planName?: string;
  status: CustomerStatus;
  balance?: number;
  ipAddress?: string;
  expiresAt?: Date;
  createdAt: Date;
}