# lib/api-zod/src/generated/types/updateCustomerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OcholaSupernet API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateCustomerBody {
  name?: string;
  phone?: string;
  email?: string;
  planId?: number;
  status?: string;
  balance?: number;
}